backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.10
PyJWT==2.8.0
Flask-Cors==4.0.0
python-dotenv==1.0.0